# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/households/invitations/inbox": {
      "get": {
        "operationId": "listInvitationInbox",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationInboxResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      }
    },
    "/api/v1/households/invitations/decline": {
      "post": {
        "operationId": "declineInvitation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "Invitation not found for this recipient."
          },
          "409": {
            "description": "Invitation is already handled, revoked, or expired."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      }
    },
    "/api/v1/households/invitations/accept": {
      "post": {
        "operationId": "acceptInvitation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetHouseholdResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invitation ID is malformed."
          },
          "404": {
            "description": "Invitation not found for this recipient."
          },
          "409": {
            "description": "Invitation is already handled, revoked, or expired."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      }
    },
    "/api/v1/households": {
      "post": {
        "operationId": "createHousehold",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseholdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateHouseholdResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Household name is invalid or the request is malformed."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      },
      "get": {
        "operationId": "listMyHouseholds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListMyHouseholdsItemDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      }
    },
    "/api/v1/households/{id}": {
      "get": {
        "operationId": "getHousehold",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetHouseholdResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Household not found or the actor is not a member."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      },
      "patch": {
        "operationId": "updateHousehold",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHouseholdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetHouseholdResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Household name is invalid or the request is malformed."
          },
          "403": {
            "description": "The actor is a member but not the current owner."
          },
          "404": {
            "description": "Household not found or the actor is not a member."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      }
    },
    "/api/v1/households/{id}/invitations": {
      "post": {
        "operationId": "sendHouseholdInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendHouseholdInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendHouseholdInvitationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Username is invalid or the request is malformed."
          },
          "403": {
            "description": "Actor is a member but not authorized to send invitations."
          },
          "404": {
            "description": "Household not found or actor is not a member."
          },
          "409": {
            "description": "The account is already a current member of this household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      },
      "get": {
        "operationId": "listInvitations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListInvitationsResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Actor is not an owner or admin of this household."
          },
          "404": {
            "description": "Household not found or actor is not a member."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      }
    },
    "/api/v1/households/{id}/invitations/{invitationId}/resend": {
      "post": {
        "operationId": "resendInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResendInvitationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invitation is already accepted, revoked, or cannot be resent."
          },
          "403": {
            "description": "Actor is not an owner or admin of this household."
          },
          "404": {
            "description": "Household or invitation not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      }
    },
    "/api/v1/households/{id}/invitations/{invitationId}/revoke": {
      "post": {
        "operationId": "revokeInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeInvitationResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Actor is not an owner or admin of this household."
          },
          "404": {
            "description": "Household or invitation not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      }
    },
    "/api/v1/households/{id}/members/{membershipId}/role": {
      "patch": {
        "operationId": "changeMemberRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetHouseholdResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Requested role is invalid or target already holds that role."
          },
          "403": {
            "description": "Actor lacks governance rights or is attempting to target the owner."
          },
          "404": {
            "description": "Household or target membership not found, or actor is not a member."
          },
          "409": {
            "description": "Stale membership state or ownership changed during the request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      }
    },
...
    "/api/v1/households/{id}/leave": {
      "post": {
        "operationId": "leaveHouseholdMembership",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Caller left; shared household content is retained."
          },
          "403": {
            "description": "OWNER_TRANSFER_REQUIRED: transfer ownership before leaving."
          },
          "404": {
            "description": "Household not found or caller is not a member."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "households"
        ]
      }
    },
...
  },
  "info": {
    "title": "Muchakucha Zwei API",
    "description": "Versioned household collaboration API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InvitationInboxResponseDto": {
        "type": "object",
        "properties": {
          "invitations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InboxInvitationDto"
            }
          }
        },
        "required": [
          "invitations"
        ]
      },
      "AcceptInvitationDto": {
        "type": "object",
        "properties": {
          "invitationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "invitationId"
        ]
      },
...
      "GetHouseholdResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "温暖小家"
          },
          "ownerMembershipId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetHouseholdMemberDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "ownerMembershipId",
          "createdAt",
          "members"
        ]
      },
...
    }
  }
}